Code refactor: change Models namespace to Domain
