+++ Resources/RothschildHouse/RothschildHouse.IdentityServer/Domain/UserClaim.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace RothschildHouse.IdentityServer.Models
+namespace RothschildHouse.IdentityServer.Domain
 {
     public class UserClaim
     {
