+++ Resources/RothschildHouse/RothschildHouse.IdentityServer/Services/ProfileService.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using IdentityServer4.Extensions;
 using IdentityServer4.Models;
 using IdentityServer4.Services;
-using RothschildHouse.IdentityServer.Models;
+using RothschildHouse.IdentityServer.Domain;
 
 namespace RothschildHouse.IdentityServer.Services
 {
