+++ Resources/RothschildHouse/RothschildHouse.IdentityServer/Startup.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using RothschildHouse.IdentityServer.Models;
+using RothschildHouse.IdentityServer.Domain;
 using RothschildHouse.IdentityServer.Services;
 using RothschildHouse.IdentityServer.Validation;
